requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
openai==1.3.7
PyPDF2==3.0.1
numpy==1.24.3
faiss-cpu==1.7.4
aiofiles==23.2.1
email-validator==2.1.0
httpx==0.24.1
